updated class diagram; removed unused code

+++ Pacman/PacmanEngine/typedefs.h
@@ -5,7 +5,7 @@
 
 namespace pac
 {
-	using ScoreType = uint64_t;
+	using ScoreType = uint32_t;
 	using TickType = uint16_t;
 	using PosType = uint16_t;
 }
